dawn_native: Add validation for binding number
The WGSL spec states that the binding number is a u16 literal. This
change adds validation for this.

See: https://gpuweb.github.io/gpuweb/wgsl/#attribute-binding
See: https://github.com/gpuweb/gpuweb/issues/1783

+++ src/dawn_native/d3d12/BindGroupLayoutD3D12.cpp
@@ -79,7 +79,7 @@ namespace dawn_native { namespace d3d12 {
 
             // TODO(dawn:728) In the future, special handling will be needed for external textures
             // here because they encompass multiple views.
-            mShaderRegisters[bindingIndex] = uint32_t(bindingInfo.binding);
+            mShaderRegisters[bindingIndex] = static_cast<uint16_t>(bindingInfo.binding);
 
             if (bindingIndex < GetDynamicBufferCount()) {
                 continue;
